+++ projectroles/email.py
@@ -251,7 +251,7 @@ def send_role_change_mail(change_type, project, user, role, request):
     :return: 1 or 0 depending on email sending success
     """
     project_url = request.build_absolute_uri(reverse(
-        'project_detail',
+        'projectroles:detail',
         kwargs={'project': project.omics_uuid}))
 
     subject = get_role_change_subject(change_type, project)