+++ samplesheets/views.py
@@ -9,10 +9,9 @@ from django.views.generic import TemplateView, FormView
 
 
 # Projectroles dependency
-from projectroles.models import Project
 from projectroles.plugins import get_backend_api
 from projectroles.views import LoggedInPermissionMixin, \
-    ProjectContextMixin, ProjectPermissionObjectMixin
+    ProjectContextMixin, ProjectPermissionMixin
 
 from .forms import SampleSheetImportForm
 from .models import Investigation, Study, Assay, Protocol, Process, \
@@ -24,8 +23,8 @@ APP_NAME = 'samplesheets'
 
 
 class ProjectSheetsView(
-        LoginRequiredMixin, LoggedInPermissionMixin, ProjectContextMixin,
-        ProjectPermissionObjectMixin, TemplateView):
+        LoginRequiredMixin, LoggedInPermissionMixin, ProjectPermissionMixin,
+        ProjectContextMixin, TemplateView):
     """Main view for displaying sample sheets in a project"""
 
     # Projectroles dependency
@@ -94,8 +93,8 @@ class ProjectSheetsView(
 
 
 class SampleSheetImportView(
-        LoginRequiredMixin, LoggedInPermissionMixin, ProjectContextMixin,
-        ProjectPermissionObjectMixin, FormView):
+        LoginRequiredMixin, LoggedInPermissionMixin, ProjectPermissionMixin,
+        ProjectContextMixin, FormView):
     """Sample sheet JSON import view"""
 
     permission_required = 'samplesheets.edit_sheet'
@@ -108,13 +107,14 @@ class SampleSheetImportView(
         kwargs = super(SampleSheetImportView, self).get_form_kwargs()
 
         if 'project' in self.kwargs:
-            kwargs.update({'project': self.kwargs['project']})
+            kwargs.update({'project': self._get_project(
+                self.kwargs, self.request)})
 
         return kwargs
 
     def form_valid(self, form):
         timeline = get_backend_api('timeline_backend')
-        project = Project.objects.get(omics_uuid=self.kwargs['project'])
+        project = self._get_project(self.kwargs, self.request)
 
         try:
             self.object = form.save()
@@ -140,20 +140,20 @@ class SampleSheetImportView(
             messages.error(self.request, str(ex))
 
         return redirect(
-            reverse('project_sheets', kwargs={
+            reverse('samplesheets:project_sheets', kwargs={
                 'project': project.omics_uuid}))
 
 
 class SampleSheetDeleteView(
         LoginRequiredMixin, LoggedInPermissionMixin, ProjectContextMixin,
-        ProjectPermissionObjectMixin, TemplateView):
+        ProjectPermissionMixin, TemplateView):
     """Sample sheet deletion view"""
     permission_required = 'samplesheets.delete_sheet'
     template_name = 'samplesheets/samplesheet_confirm_delete.html'
 
     def post(self, request, *args, **kwargs):
         timeline = get_backend_api('timeline_backend')
-        project = Project.objects.get(omics_uuid=self.kwargs['project'])
+        project = self._get_project(request, kwargs)
         investigation = Investigation.objects.get(project=project)
 
         # Add event in Timeline
@@ -176,5 +176,6 @@ class SampleSheetDeleteView(
         messages.success(
             self.request, 'Sample sheets deleted.')
 
-        return HttpResponseRedirect(reverse('project_sheets', kwargs={
-            'project': project.omics_uuid}))
+        return HttpResponseRedirect(reverse(
+            'samplesheets:project_sheets',
+            kwargs={'project': project.omics_uuid}))