+++ timeline/api.py
@@ -112,7 +112,7 @@ class TimelineAPI:
                     label=r)
 
                 # Get history link
-                history_url = reverse('object_timeline', kwargs={
+                history_url = reverse('timeline:object_timeline', kwargs={
                     'project': event.project.omics_uuid,
                     'object_model': ref_obj.object_model,
                     'object_uuid': ref_obj.object_uuid})
@@ -140,7 +140,7 @@ class TimelineAPI:
                                 'projectroles.view_project', project):
                             refs[r] = '<a href="{}">{}</a>'.format(
                                 reverse(
-                                    'project_detail',
+                                    'projetroles:detail',
                                     kwargs={'project': project.omics_uuid}),
                                 project.title)
 
@@ -190,7 +190,7 @@ class TimelineAPI:
         :param obj: Django postgres database object
         :return: String
         """
-        return reverse('object_timeline', kwargs={
+        return reverse('timeline:object_timeline', kwargs={
             'project': project_uuid,
             'object_model': obj.__class__.__name__,
             'object_uuid': obj.omics_uuid})