+++ timeline/plugins.py
@@ -49,7 +49,7 @@ class ProjectAppPlugin(ProjectAppPluginPoint):
     icon = 'clock-o'
 
     #: Entry point URL ID (must take project omics_uuid as "project" argument)
-    entry_point_url_id = 'project_timeline'
+    entry_point_url_id = 'timeline:project_timeline'
 
     #: Description string
     description = 'Timeline of project events'